+++ apps/portal/src/components/Layouts/DocLayout.tsx
@@ -57,12 +57,7 @@ export function DocLayout(props: DocLayoutProps) {
         data-no-llm={props.noLLM}
         data-noindex={props.noIndex}
       >
-        <div className="grow xl:mt-6">
-          <h5 className="mb-2 text-sm text-muted-foreground">
-            {props.sideBar.name}
-          </h5>
-          {props.children}
-        </div>
+        <div className="grow xl:mt-6">{props.children}</div>
         <div className="mt-16 xl:mt-20">
           <PageFooter
             editPageButton={props.editPageButton}