+++ apps/portal/src/components/others/Sidebar.tsx
@@ -51,18 +51,28 @@ type ReferenceSideBarProps = {
   onLinkClick?: () => void;
   header?: React.ReactNode;
   name: string;
+  className?: string;
 };
 
 export function DocSidebar(props: ReferenceSideBarProps) {
   return (
-    <div className="flex h-full flex-col pb-10 pt-6 text-muted-foreground text-sm">
+    <div
+      className={cn(
+        "flex h-full flex-col pb-10 pt-6 text-muted-foreground text-sm",
+        props.className,
+      )}
+    >
       {/* Side bar Name */}
       {props.header}
       <ul className="styled-scrollbar transform-gpu space-y-1">
         {props.links.map((link, i) => (
           // biome-ignore lint/suspicious/noArrayIndexKey: TODO - fix this
           <li key={i}>
-            <SidebarItem link={link} onLinkClick={props.onLinkClick} />
+            <SidebarItem
+              link={link}
+              onLinkClick={props.onLinkClick}
+              isFirst={i === 0}
+            />
           </li>
         ))}
       </ul>
@@ -70,7 +80,11 @@ export function DocSidebar(props: ReferenceSideBarProps) {
   );
 }
 
-function SidebarItem(props: { link: SidebarLink; onLinkClick?: () => void }) {
+function SidebarItem(props: {
+  link: SidebarLink;
+  onLinkClick?: () => void;
+  isFirst: boolean;
+}) {
   const pathname = usePathname();
 
   if ("separator" in props.link) {
@@ -86,6 +100,7 @@ function SidebarItem(props: { link: SidebarLink; onLinkClick?: () => void }) {
     if (link.isCollapsible === false) {
       return (
         <DocSidebarNonCollapsible
+          isFirst={props.isFirst}
           key={link.name}
           linkGroup={link}
           onLinkClick={props.onLinkClick}
@@ -139,13 +154,14 @@ function SidebarItem(props: { link: SidebarLink; onLinkClick?: () => void }) {
 function DocSidebarNonCollapsible(props: {
   linkGroup: LinkGroup;
   onLinkClick?: () => void;
+  isFirst: boolean;
 }) {
   const pathname = usePathname();
   const { href, name, links, icon } = props.linkGroup;
   const isCategoryActive = href ? isSamePage(pathname, href) : false;
 
   return (
-    <div className="my-4">
+    <div className={cn("my-4", props.isFirst && "mt-0")}>
       <div className="mb-2 flex items-center gap-2 rounded-lg text-foreground">
         {icon && <SidebarIcon icon={icon} />}
         {href ? (
@@ -167,7 +183,11 @@ function DocSidebarNonCollapsible(props: {
           return (
             // biome-ignore lint/suspicious/noArrayIndexKey: TODO - fix this
             <li key={i}>
-              <SidebarItem link={link} onLinkClick={props.onLinkClick} />
+              <SidebarItem
+                link={link}
+                onLinkClick={props.onLinkClick}
+                isFirst={i === 0}
+              />
             </li>
           );
         })}
@@ -232,7 +252,11 @@ function DocSidebarCategory(props: {
             return (
               // biome-ignore lint/suspicious/noArrayIndexKey: TODO - fix this
               <li key={i}>
-                <SidebarItem link={link} onLinkClick={props.onLinkClick} />
+                <SidebarItem
+                  link={link}
+                  onLinkClick={props.onLinkClick}
+                  isFirst={i === 0}
+                />
               </li>
             );
           })}