+++ packages/thirdweb/src/react/web/ui/ConnectWallet/useConnectModal.tsx
@@ -118,6 +118,7 @@ function Modal(
       termsOfServiceUrl: props.termsOfServiceUrl,
       title: props.title,
       titleIconUrl: props.titleIcon,
+      requireApproval: props.requireApproval,
     };
   }, [
     props.privacyPolicyUrl,
@@ -125,6 +126,7 @@ function Modal(
     props.termsOfServiceUrl,
     props.title,
     props.titleIcon,
+    props.requireApproval,
   ]);
 
   return (
@@ -449,6 +451,22 @@ export type UseConnectModalOptions = {
    * Refer to the [`SiweAuthOptions`](https://portal.thirdweb.com/references/typescript/v5/SiweAuthOptions) for more details
    */
   auth?: SiweAuthOptions;
+
+  /**
+   * Require terms of service and privacy policy to be accepted before connecting an in-app wallet.
+   *
+   * By default it's `false`
+   * @example
+   * ```tsx
+   * function Example() {
+   *   const { connect } = useConnectModal();
+   *   return <button onClick={() => connect({ client, requireApproval: true })}>
+   *     Connect
+   *   </button>
+   * }
+   * ```
+   */
+  requireApproval?: boolean;
 };
 
 // TODO: consilidate Button/Embed/Modal props into one type with extras