+++ apps/wagmi-demo/CHANGELOG.md
@@ -1,5 +1,29 @@
 # wagmi-inapp
 
+## 0.0.52
+
+### Patch Changes
+
+- Updated dependencies [[`5352152`](https://github.com/thirdweb-dev/js/commit/53521528353a00691e82405460d4fb869ac986de)]:
+  - thirdweb@5.119.0
+  - @thirdweb-dev/wagmi-adapter@0.2.207
+
+## 0.0.51
+
+### Patch Changes
+
+- Updated dependencies [[`1caf6dd`](https://github.com/thirdweb-dev/js/commit/1caf6dd4609ac72a06999ef0d70f31a371e5108a)]:
+  - thirdweb@5.118.2
+  - @thirdweb-dev/wagmi-adapter@0.2.206
+
+## 0.0.50
+
+### Patch Changes
+
+- Updated dependencies [[`b50e304`](https://github.com/thirdweb-dev/js/commit/b50e304e0ccd1ac7eae551c3d49e82312ffc82d1)]:
+  - thirdweb@5.118.1
+  - @thirdweb-dev/wagmi-adapter@0.2.205
+
 ## 0.0.49
 
 ### Patch Changes